fix(runtime): restore authenticated chat and modularize settings
